feat: add platformVersion
